Reorder toolbar views methods
